refactror mcp
