+++ mcp_tools/web_app_security/additional_security_tools.py
@@ -5,70 +5,6 @@ from typing import Dict, Any
 def register_additional_security_tools(mcp, hexstrike_client, logger):
 
     @mcp.tool()
-    def ffuf_scan(url: str, wordlist: str = "/usr/share/wordlists/dirb/common.txt", mode: str = "directory", match_codes: str = "200,204,301,302,307,401,403", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute FFuf for web fuzzing with enhanced logging.
-
-        Args:
-            url: The target URL
-            wordlist: Wordlist file to use
-            mode: Fuzzing mode (directory, vhost, parameter)
-            match_codes: HTTP status codes to match
-            additional_args: Additional FFuf arguments
-
-        Returns:
-            Web fuzzing results
-        """
-        data = {
-            "url": url,
-            "wordlist": wordlist,
-            "mode": mode,
-            "match_codes": match_codes,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting FFuf {mode} fuzzing: {url}")
-        result = hexstrike_client.safe_post("api/tools/ffuf", data)
-        if result.get("success"):
-            logger.info(f"✅ FFuf fuzzing completed for {url}")
-        else:
-            logger.error(f"❌ FFuf fuzzing failed for {url}")
-        return result
-
-    @mcp.tool()
-    def netexec_scan(target: str, protocol: str = "smb", username: str = "", password: str = "", hash_value: str = "", module: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute NetExec (formerly CrackMapExec) for network enumeration with enhanced logging.
-
-        Args:
-            target: The target IP or network
-            protocol: Protocol to use (smb, ssh, winrm, etc.)
-            username: Username for authentication
-            password: Password for authentication
-            hash_value: Hash for pass-the-hash attacks
-            module: NetExec module to execute
-            additional_args: Additional NetExec arguments
-
-        Returns:
-            Network enumeration results
-        """
-        data = {
-            "target": target,
-            "protocol": protocol,
-            "username": username,
-            "password": password,
-            "hash": hash_value,
-            "module": module,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting NetExec {protocol} scan: {target}")
-        result = hexstrike_client.safe_post("api/tools/netexec", data)
-        if result.get("success"):
-            logger.info(f"✅ NetExec scan completed for {target}")
-        else:
-            logger.error(f"❌ NetExec scan failed for {target}")
-        return result
-
-    @mcp.tool()
     def amass_scan(domain: str, mode: str = "enum", additional_args: str = "") -> Dict[str, Any]:
         """
         Execute Amass for subdomain enumeration with enhanced logging.
