+++ mcp_tools/web_fuzz/ffuf.py
@@ -1,0 +1,34 @@
+# mcp_tools/web_fuzz/ffuf.py
+
+from typing import Dict, Any
+
+def register_ffuf_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def ffuf_scan(url: str, wordlist: str = "/usr/share/wordlists/dirb/common.txt", mode: str = "directory", match_codes: str = "200,204,301,302,307,401,403", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute FFuf for web fuzzing with enhanced logging.
+
+        Args:
+            url: The target URL
+            wordlist: Wordlist file to use
+            mode: Fuzzing mode (directory, vhost, parameter)
+            match_codes: HTTP status codes to match
+            additional_args: Additional FFuf arguments
+
+        Returns:
+            Web fuzzing results
+        """
+        data = {
+            "url": url,
+            "wordlist": wordlist,
+            "mode": mode,
+            "match_codes": match_codes,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting FFuf {mode} fuzzing: {url}")
+        result = hexstrike_client.safe_post("api/tools/ffuf", data)
+        if result.get("success"):
+            logger.info(f"✅ FFuf fuzzing completed for {url}")
+        else:
+            logger.error(f"❌ FFuf fuzzing failed for {url}")
+        return result
